Crate collectives_polkadot_integration_tests

Re-exports

pub use xcm::v3::NetworkId::Polkadot as PolkadotId;

pub use xcm_emulator::cumulus_pallet_dmp_queue;

pub use xcm_emulator::AccountId32Junction;

Modules

opaque

Macros

assert_err

Assert an expression returns an error specified.

assert_expected_events

assert_ok

Panic if an expression doesn’t evaluate to Ok.

bx

lazy_static

Structs

AccountId32

An opaque 32-byte cryptographic identifier.

AlwaysV2

WrapVersion implementation which attempts to always convert the XCM to version 2 before wrapping it.

AlwaysV3

WrapVersion implementation which attempts to always convert the XCM to version 3 before wrapping it.

Ancestor

A unit struct which can be converted into a MultiLocation of the inner parents value.

AncestorThen

A unit struct which can be converted into a MultiLocation of the inner parents value and the inner interior.

AssetHubPolkadot

AssetHubPolkadotReceiver

AssetHubPolkadotSender

BridgeHubPolkadot

BridgeHubPolkadotReceiver

BridgeHubPolkadotSender

Collectives

CollectivesReceiver

CollectivesSender

DoubleEncoded

Wrapper around the encoded and decoded versions of a value. Caches the decoded value once computed.

HrmpChannelId

A type that uniquely identifies an HRMP channel. An HRMP channel is established between two paras. In text, we use the notation (A, B) to specify a channel between A and B. The channels are unidirectional, meaning that (A, B) and (B, A) refer to different channels. The convention is that we use the first item tuple for the sender and the second for the recipient. Only one channel is allowed between two participants in one direction, i.e. there cannot be 2 different channels identified by (A, B). A channel with the same para id in sender and recipient is invalid. That is, however, not enforced.

Id

Unique identifier of a parachain.

InboundDownwardMessage

A wrapped version of DownwardMessage. The difference is that it has attached the block number when the message was sent.

Instance1

Instance1 to be used for instantiable palllets defined with the #[pallet] macro. Instances 2-16 are also available but are hidden from docs.

MultiAsset

Either an amount of a single fungible asset, or a single well-identified non-fungible asset.

MultiAssets

A Vec of MultiAssets.

MultiLocation

A relative path between state-bearing consensus systems.

PalletInfo

ParaId

Unique identifier of a parachain.

Parent

A unit struct which can be converted into a MultiLocation of parents value 1.

ParentThen

A tuple struct which can be converted into a MultiLocation of parents value 1 with the inner interior.

PenpalPolkadotA

PenpalPolkadotAReceiver

PenpalPolkadotASender

PenpalPolkadotB

PenpalPolkadotBReceiver

PenpalPolkadotBSender

Polkadot

PolkadotMockNet

PolkadotReceiver

PolkadotSender

QueryResponseInfo

Information regarding the composition of a query response.

Test

Struct that help with tests where either dispatchables or assertions need to be reused. The struct keeps the test’s arguments of your choice in the generic Args. These arguments can be easily reused and shared between the assertions functions and dispatchables functions, which are also stored in Test. Origin corresponds to the chain where the XCM interaction starts with an initial execution. Destination corresponds to the last chain where an effect of the intial execution is expected happen. Hops refer all the ordered intermediary chains an initial XCM execution can provoke some effect.

TestArgs

Default Args provided by xcm-emulator to be stored in a Test instance

TestContext

Auxiliar struct to help creating a new Test instance

Weight

Xcm

XcmContext

Contextual data pertaining to a specific list of XCM instructions.

Enums

AggregateMessageOrigin

Aggregate message origin for the MessageQueue pallet.

AssetId

Classification of an asset being concrete or abstract.

AssetInstance

A general identifier for an instance of a non-fungible asset class.

BodyId

An identifier of a pluralistic body.

BodyPart

A part of a pluralistic body.

DispatchError

Reason why a dispatch call failed.

Error

Error codes used in XCM. The first errors codes have explicit indices and are part of the XCM format. Those trailing are merely part of the XCM implementation; there is no expectation that they will retain the same index over time.

Fungibility

Classification of whether an asset is fungible or not, along with a mandatory amount or instance.

Instruction

Cross-Consensus Message: A message from one consensus system to another.

Junction

A single item in a path to describe the relative location of a consensus system.

Junctions

Non-parent junctions that can be constructed, up to the length of 8. This specific Junctions implementation uses a Rust enum in order to make pattern matching easier.

MaybeErrorCode

MultiAddress

A multi-format address wrapper for on-chain accounts.

MultiAssetFilter

MultiAsset collection, defined either by a number of MultiAssets or a single wildcard.

NetworkId

A global identifier of a data structure existing within consensus.

OriginKind

Basically just the XCM (more general) version of ParachainDispatchOrigin.

Outcome

Outcome of an XCM execution.

Response

Response data to a query.

SendError

Error result value when attempting to send an XCM message.

UmpQueueId

Identifies a UMP queue inside the MessageQueue pallet.

Unsupported

VersionedAssetId

A single version’s Response value, together with its version code.

VersionedInteriorMultiLocation

A single InteriorMultiLocation value, together with its version code.

VersionedMultiAsset

A single MultiAsset value, together with its version code.

VersionedMultiAssets

A single MultiAssets value, together with its version code.

VersionedMultiLocation

A single MultiLocation value, together with its version code.

VersionedResponse

A single version’s Response value, together with its version code.

VersionedXcm

A single XCM message, together with its version code.

WeightLimit

An optional weight limit.

WildFungibility

Classification of whether an asset is fungible or not.

WildMultiAsset

A wildcard representing a set of assets.

XcmError

Error codes used in XCM. The first errors codes have explicit indices and are part of the XCM format. Those trailing are merely part of the XCM implementation; there is no expectation that they will retain the same index over time.

Constants

ALICE

ASSETS_PALLET_ID

ASSET_HUB_POLKADOT_ED

ASSET_ID

ASSET_MIN_BALANCE

BOB

POLKADOT_ED

PROOF_SIZE_THRESHOLD

REF_TIME_THRESHOLD

XCM_V3

XCM_VERSION

This module’s XCM version.

Traits

AssetHubPolkadotPallet

BridgeHubPolkadotPallet

Chain

CollectivesPallet

Encode

Trait that allows zero-copy write of value-references to slices in LE format.

ExecuteXcm

Type of XCM message executor.

Inspect

Trait for providing balance-inspection access to a set of named fungible assets.

IntoVersion

Attempt to convert self into a particular version of itself.

OriginTrait

Methods available on frame_system::Config::RuntimeOrigin.

Para

PenpalPolkadotAPallet

PenpalPolkadotBPallet

PolkadotPallet

PreparedMessage

Relay

SendXcm

Utility for sending an XCM message to a given location.

TestExt

WrapVersion

Convert an Xcm datum into a VersionedXcm, based on a destination MultiLocation which will interpret it.

XcmWeightInfo

Functions

relay_test_args

Returns a TestArgs instance to de used for the Relay Chain accross integraton tests

send_xcm

Convenience function for using a SendXcm implementation. Just interprets the dest and wraps both in Some before passing them as as mutable references into T::send_xcm.

system_para_test_args

Returns a TestArgs instance to de used for the System Parachain accross integraton tests

validate_send

Convenience function for using a SendXcm implementation. Just interprets the dest and wraps both in Some before passing them as as mutable references into T::send_xcm.

weight_within_threshold

xcm_transact_paid_execution

Helper method to build a XCM with a Transact instruction and paying for its execution

xcm_transact_unpaid_execution

Helper method to build a XCM with a Transact instruction without paying for its execution

Type Aliases

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AlwaysLatest

WrapVersion implementation which attempts to always convert the XCM to the latest version before wrapping it.

AlwaysLts

WrapVersion implementation which attempts to always convert the XCM to the most recent Long- Term-Support version before wrapping it.

Balance

Balance of an account.

DispatchResult

This is the legacy return type of Dispatchable. It is still exposed for compatibility reasons. The new return type is DispatchResultWithInfo. FRAME runtimes should use frame_support::dispatch::DispatchResult.

InteriorMultiLocation

A relative location which is constrained to be an interior location of the context.

QueryId

An identifier for a query.

RelayToSystemParaTest

SendResult

Result value when attempting to send an XCM message.

SystemParaToParaTest

SystemParaToRelayTest

TestExternalities

Type alias for Externalities implementation used in tests.

XcmHash

A hash type for identifying messages.

XcmResult

XcmVersion

A version of XCM.

Derive Macros

Encode

Derive parity_scale_codec::Encode and parity_scale_codec::EncodeLike for struct and enum.